+++ vortex-tensor/src/scalar_fns/l2_norm.rs
@@ -48,6 +48,7 @@ use vortex_session::VortexSession;
 use crate::matcher::AnyTensor;
 use crate::scalar_fns::l2_denorm::NormalForm;
 use crate::utils::extract_flat_elements;
+use crate::utils::inner_vector_array;
 use crate::utils::validate_tensor_float_input;
 
 /// L2 norm (Euclidean norm) of a tensor or vector column.
@@ -186,7 +187,7 @@ impl ScalarFnVTable for L2Norm {
 
         // Drill past any `NormalizedVector` wrapper so we always work with the underlying
         // `Vector` extension array.
-        let input_ref = crate::types::normalized_vector::inner_vector_array(&input_ref, ctx)?;
+        let input_ref = inner_vector_array(&input_ref, ctx)?;
         let input: ExtensionArray = input_ref.execute(ctx)?;
         let validity = input.as_ref().validity()?;
 
@@ -287,7 +288,7 @@ fn execute_normalized_vector_norms(
 ) -> VortexResult<ArrayRef> {
     // `NormalizedVector` storage is `Extension(Vector(FSL))`; drill to the inner `Vector` to
     // reach the underlying FSL.
-    let vector_ref = crate::types::normalized_vector::inner_vector_array(input_ref, ctx)?;
+    let vector_ref = inner_vector_array(input_ref, ctx)?;
     let input: ExtensionArray = vector_ref.execute(ctx)?;
     let validity = input.as_ref().validity()?;
     let flat = extract_flat_elements(input.storage_array(), tensor_flat_size, ctx)?;
